requirements.txt
Django==1.11.22
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
